+++ src/main/java/com/shop/controller/HomeController.java
@@ -33,40 +33,12 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class HomeController {
 
-	@Setter(onMethod_ = @Autowired)
-	private CarShopService service;
-
 	private static final Logger logger = LoggerFactory.getLogger(HomeController.class);
 
 	/**
 	 * Simply selects the home view to render by returning its name.
 	 */
 
-	@RequestMapping({ "/", "index" })
-	public String index(Model model) {
-		model.addAttribute("list", service.index());
-
-		return "carshop/index";
-
-	}
-
-
-	@RequestMapping("/checkout")
-	public String checkout() {
-		return "carshop/checkout";
-	}
-
-	
-	@RequestMapping("/like")
-	public String like() {
-		return"carshop/like";
-	}	
-	
-	@RequestMapping("/cart")
-	public String cart() {
-		return"carshop/cart";
-	}	
-	
 	
 	@RequestMapping("/fake")
 	public String fake() {
@@ -78,33 +50,6 @@ public class HomeController {
 		return "home";
 	}
 
-
-	@RequestMapping("/mypage")
-	public String mypage() {
-		return "carshop/mypage";
-		
-	}
-
-	@GetMapping("/productForm")
-	public String productForm(Model model) {
-		System.out.println("productForm 들어옴");
-		model.addAttribute("cateParent",service.cateParent());
-		model.addAttribute("category", service.category());
-		return "carshop/productForm";
-	}
-
-	//action 에서 post 방식으로 들어올 시 입력 폼 
-	@PostMapping("/productForm")
-	public String register(Prod_Tbl product) {
-		log.info("컨틀롤러에서 등록 " + product);
-		service.productForm(product);
-		return "redirect:/carshop/index";
-	}
-
-	@RequestMapping("/product/details")
-	public String detail() {
-		return "carshop/productdetails";
-	}
 	
 
 }
